docs(text): added more doc comments for texts
